refactor: src/ renamed to marsbench/
