+++ configs/model/classification/resnet18.yaml
@@ -1,5 +1,5 @@
 name: ResNet18
-class_path: src.models.classification.ResNet18
+class_path: marsbench.models.classification.ResNet18
 status: ready
 pretrained: true
 freeze_layers: true
