+++ configs/model/classification/resnet50.yaml
@@ -1,5 +1,5 @@
 name: ResNet50
-class_path: src.models.classification.ResNet50
+class_path: marsbench.models.classification.ResNet50
 status: ready
 pretrained: true
 freeze_layers: true
