+++ configs/model/classification/squeezenet.yaml
@@ -1,5 +1,5 @@
 name: SqueezeNet
-class_path: src.models.classification.SqueezeNet
+class_path: marsbench.models.classification.SqueezeNet
 status: ready
 pretrained: true
 freeze_layers: true
