+++ configs/model/classification/swin_transformer.yaml
@@ -1,5 +1,5 @@
 name: SwinTransformer
-class_path: src.models.classification.SwinTransformer
+class_path: marsbench.models.classification.SwinTransformer
 status: ready
 pretrained: true
 freeze_layers: true
