+++ configs/model/segmentation/unet.yaml
@@ -1,5 +1,5 @@
 name: UNet
-class_path: src.models.segmentation.UNet
+class_path: marsbench.models.segmentation.UNet
 status: test
 pretrained: true
 freeze_layers: true
