+++ marsbench/models/classification/SqueezeNet.py
